Adding the start of GC, and the Tri REI and RSymbol classes. TriREI covers true/nil/false and the instantiation of their classes. RSymbol is the thing we pass around when we talk about a Symbol. The idea is still experimental. Notes about RubyValue follows.
RubyValue is now a horrible union-based thing, as I want to use it as a value type (which creates problems when we want to use subclasses like we were before). RubySymbol is one of the union members, so we can track its existence (from within the Environment, as before)

This commit is kinda out of order.